requirements.txt
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bcrypt==3.1.7
bleach==3.2.1
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
coverage==5.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.2
django-filter==2.4.0
djangorestframework==3.12.2
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
httpie==2.3.0
httplib2==0.14.0
idna==2.8
iniconfig==1.0.1
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.5.4
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.3
MarkupSafe==1.1.0
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
netifaces==0.10.4
nose2==0.9.2
notebook==6.1.4
oauthlib==3.1.0
olefile==0.46
packaging==20.4
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pipreqs==0.4.10
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.6.1
ptyprocess==0.6.0
py==1.9.0
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
Pygments==2.7.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
pymongo==3.11.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.17.3
PySocks==1.7.1
pytest==6.1.1
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
reportlab==3.5.34
requests==2.22.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.1
systemd-python==234
terminado==0.9.1
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==5.0.4
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.34
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
yarg==0.1.9
